pyproject.toml
[project]
name = "mitmproxy-openai"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = []

[dependency-groups]
dev = [
    "jinja2>=3.1.6",
    "mitmproxy>=12.2.0",
    "pydantic>=2.12.4",
    "pyyaml>=6.0.3",
]
